+++ engine/lite/components/kafka-tests/src/test/scala/pl/touk/nussknacker/engine/lite/components/LiteKafkaUniversalJsonFunctionalTest.scala
@@ -1,6 +1,6 @@
 package pl.touk.nussknacker.engine.lite.components
 
-import cats.data.Validated.{Invalid, Valid}
+import cats.data.Validated.{Invalid, Valid, invalidNel}
 import cats.data.{NonEmptyList, Validated, ValidatedNel}
 import io.circe.Json
 import io.circe.Json.{Null, fromFields, fromInt, fromJsonObject, fromLong, fromString, obj}
@@ -170,21 +170,21 @@ class LiteKafkaUniversalJsonFunctionalTest
       (sampleMapAny,      schemaMapAny,                  schemaMapStr,                         strict,             invalidTypes("path 'value' actual: 'Unknown' expected: 'String'")),
       (sampleMapStr,      schemaMapAny,                  schemaMapStr,                         lax,                valid(sampleMapStr)),
       (sampleMapStr,      schemaMapStr,                  schemaMapStr,                         strictAndLax,       valid(sampleMapStr)),
-      (sampleMapStr,      schemaMapStringOrLong,         schemaMapStr,                         strict,             invalidTypes("path 'value' actual: 'String | Long' expected: 'String'")),
+      (sampleMapStr,      schemaMapStringOrLong,         schemaMapStr,                         strict,             invalidTypes("path 'value' actual: 'Long | String' expected: 'String'")),
       (sampleMapStr,      schemaMapStringOrLong,         schemaMapStr,                         lax,                valid(sampleMapStr)),
       (sampleMapPerson,   schemaMapObjPerson,            schemaMapStr,                         strictAndLax,       invalidTypes("path 'value' actual: 'Record{age: Long, first: String, last: String}' expected: 'String'")),
       (sampleMapPerson,   schemaMapObjPersonWithLimits,  schemaMapStr,                         strictAndLax,       invalidTypes("path 'value' actual: 'Record{age: Integer, first: String, last: String}' expected: 'String'")),
       (sampleArrayInt,    schemaArrayLong,               schemaMapStr,                         strictAndLax,       invalidTypes("actual: 'List[Long]' expected: 'Map[String, String]'")),
       (samplePerson,      schemaPerson,                  schemaMapStr,                         strictAndLax,       invalidTypes("path 'age' actual: 'Long' expected: 'String'")),
       (samplePerson,      schemaPersonWithLimits,        schemaMapStr,                         strictAndLax,       invalidTypes("path 'age' actual: 'Integer' expected: 'String'")),
-      (sampleMapAny,      schemaMapAny,                  schemaMapStringOrLong,                strict,             invalidTypes("path 'value' actual: 'Unknown' expected: 'String | Long'")),
+      (sampleMapAny,      schemaMapAny,                  schemaMapStringOrLong,                strict,             invalidTypes("path 'value' actual: 'Unknown' expected: 'Long | String'")),
       (sampleMapInt,      schemaMapAny,                  schemaMapStringOrLong,                lax,                valid(sampleMapInt)),
       (sampleMapStr,      schemaMapStr,                  schemaMapStringOrLong,                strictAndLax,       valid(sampleMapStr)),
       (sampleMapStr,      schemaMapStringOrLong,         schemaMapStringOrLong,                strictAndLax,       valid(sampleMapStr)),
       (sampleMapInt,      schemaMapStringOrLong,         schemaMapStringOrLong,                strictAndLax,       valid(sampleMapInt)),
-      (samplePerson,      schemaMapObjPerson,            schemaMapStringOrLong,                strictAndLax,       invalidTypes("path 'value' actual: 'Record{age: Long, first: String, last: String}' expected: 'String | Long'")),
-      (samplePerson,      schemaMapObjPersonWithLimits,  schemaMapStringOrLong,                strictAndLax,       invalidTypes("path 'value' actual: 'Record{age: Integer, first: String, last: String}' expected: 'String | Long'")),
-      (sampleArrayInt,    schemaArrayLong,               schemaMapStringOrLong,                strictAndLax,       invalidTypes("actual: 'List[Long]' expected: 'Map[String, String | Long]'")),
+      (samplePerson,      schemaMapObjPerson,            schemaMapStringOrLong,                strictAndLax,       invalidTypes("path 'value' actual: 'Record{age: Long, first: String, last: String}' expected: 'Long | String'")),
+      (samplePerson,      schemaMapObjPersonWithLimits,  schemaMapStringOrLong,                strictAndLax,       invalidTypes("path 'value' actual: 'Record{age: Integer, first: String, last: String}' expected: 'Long | String'")),
+      (sampleArrayInt,    schemaArrayLong,               schemaMapStringOrLong,                strictAndLax,       invalidTypes("actual: 'List[Long]' expected: 'Map[String, Long | String]'")),
       (samplePerson,      schemaPerson,                  schemaMapStringOrLong,                strictAndLax,       valid(samplePerson)),
       (samplePerson,      schemaPerson,                  nameAndLastNameSchema,                strictAndLax,       valid(samplePerson)),
       (samplePerson,      schemaPerson,                  nameAndLastNameSchema(schemaLong),    strictAndLax,       valid(samplePerson)),
@@ -252,9 +252,9 @@ class LiteKafkaUniversalJsonFunctionalTest
       (A,         schemaString,             schemaEnumABC,         strictAndLax,       valid(A)),
       (A,         schemaEnumABC,            schemaString,          strictAndLax,       valid(A)),
       (A,         schemaEnumABC,            schemaEnumAB1,         lax,                valid(A)),
-      (A,         schemaEnumABC,            schemaEnumAB1,         strict,             invalidTypes("actual: 'String(A) | String(B) | String(C)' expected: 'String(A) | Integer(1) | String(B)'")),
+      (A,         schemaEnumABC,            schemaEnumAB1,         strict,             invalidTypes("actual: 'String(A) | String(B) | String(C)' expected: 'Integer(1) | String(A) | String(B)'")),
       (A,         schemaEnumAB1,            schemaEnumAB,          lax,                valid(A)),
-      (A,         schemaEnumAB1,            schemaEnumAB,          strict,             invalidTypes("actual: 'String(A) | Integer(1) | String(B)' expected: 'String(A) | String(B)'")),
+      (A,         schemaEnumAB1,            schemaEnumAB,          strict,             invalidTypes("actual: 'Integer(1) | String(A) | String(B)' expected: 'String(A) | String(B)'")),
       (A,         schemaEnumAB1,            schemaEnumAB1,         strictAndLax,       valid(A)),
       (one,       schemaEnumAB1,            schemaEnumAB1,         strictAndLax,       valid(one)),
       (obj,       schemaEnumStrOrObj,       schemaEnumStrOrObj,    lax,                valid(obj)),
@@ -319,7 +319,7 @@ class LiteKafkaUniversalJsonFunctionalTest
       (inputObject,                  objWithPatternPropsAndStringAdditionalSchema,         objWithPatternPropsAndStringAdditionalSchema,   Input,                                        strict,               invalidTypes("actual: 'Map[String,Long | String]' expected: 'Map[String, String]'")),
       (inputObject,                  objWithPatternPropsAndStringAdditionalSchema,         schemaLong,                                     SpecialSpELElement("#input['foo_int']"),      lax,                  valid(inputObjectIntPropValue)),
       (inputObject,                  objWithPatternPropsAndStringAdditionalSchema,         schemaLong,                                     SpecialSpELElement("#input['foo_int']"),      strict,               invalidTypes("actual: 'Long | String' expected: 'Long'")),
-      (inputObject,                  objWithDefinedPropsPatternPropsAndAdditionalSchema,   schemaLong,                                     SpecialSpELElement("#input['foo_int']"),      lax,                  Invalid(NonEmptyList(ExpressionParserCompilationError("There is no property 'foo_int' in type: Record{definedProp: String}", "my-sink", Some("Value"), "#input['foo_int']"), Nil))),
+      (inputObject,                  objWithDefinedPropsPatternPropsAndAdditionalSchema,   schemaLong,                                     SpecialSpELElement("#input['foo_int']"),      lax,                  invalidNel(ExpressionParserCompilationError("There is no property 'foo_int' in type: Record{definedProp: String}", "my-sink", Some("Value"), "#input['foo_int']"))),
       (inputObjectWithDefinedProp,   objWithDefinedPropsPatternPropsAndAdditionalSchema,   schemaString,                                   SpecialSpELElement("#input.definedProp"),     strict,               valid(inputObjectDefinedPropValue)),
     )
     //@formatter:on